\name{contagion}
\alias{contagion}
\docType{data}
\title{Guatemalen Women's Perceptions of Contagion in Disease}
\description{This data set was collected by Weller (1983,1984) and used in Romney, et al. (1986).  Twenty-four urban Guatemalen women were asked about the role of contagion in 27 diseases.
}
\usage{data(contagion)}
\format{
There are two objects: 

questions: list of diseases 

answermat: 24 by 27 matrix of responses
}
\source{
Susan C. Weller (1983) New Data on Intracultural Variability: The Hot-cold Concept of Medicine and Illness. \emph{Human Organization} \bold{42(3)}: 249-257.
}
\references{
Kimball A. Romney, Susan C. Weller, and William H. Batchelder (1986).  Culture as Consensus: A Theory of Culture and Informant Accuract.  \emph{American Anthropologist} \bold{88(2)}

Susan C. Weller (1984) Cross Cultural Concepts of Illness: Variation and Validation.  \emph{American Anthropologist} \bold{86(2)}:341-351.
}
\examples{
data(contagion)
contagion$questions
contagion$answermat
}
\keyword{datasets}
