\name{SubCultCon-package}
\alias{SubCultCon-package}
\alias{SubCultCon}
\docType{package}
\title{
Subcultural Consensus Modeling
}
\description{
The three functions in the package compute the maximum likelihood estimates of the informants' competence scores, test for two answer keys with known groups, and find best split of the informants into sub-culture groups.
}
\details{
\tabular{ll}{
Package: \tab SubCultCon\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-09-25\cr
License: \tab GPL-2 or GPL-3\cr
}
}
\author{
Mary C Meyer, Jeffrey G Snodgrass, Michael Lacy

Maintainer: Mary C Meyer <meyer@stat.colostate.edu>
}
\references{
Kimball A.	Romney, Susan C. Weller, and William H. Batchelder (1986).  Culture as Consensus: A Theory of Culture and Informant Accuracy.  \emph{American Anthropologist}, \bold{88(2)}, 313-338.
}
\keyword{ package }
\seealso{
\code{\link{ccgrouptest}}, \code{\link{GAmaxcomp}}, \code{\link{ccmle}}, \code{\link{contagion}}, , \code{\link{village}}
}
